rework cli overview
