+++ lasr/src/lasrctl/mod.rs
@@ -1,4 +1,3 @@
 pub mod builders;
 pub mod cli;
-pub mod commands;
 pub mod network;
